requirements.txt
discord
python-dotenv==1.0.0